pyproject.toml
[project]
name = "cine-reserve"
version = "0.1.0"
description = ""
authors = [
    {name = "afredoherculano",email = "alfredoherculano@proton.me"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
